rename player folder
